+++ packages/react-spinbutton/src/SpinButton.ts
@@ -1,0 +1,1 @@
+export * from './components/SpinButton/index';
