+++ packages/react-spinbutton/src/components/SpinButton/examples/SpinButton.BasicWithIcon.Example.tsx
