+++ packages/react-spinbutton/src/index.ts
@@ -1,1 +1,1 @@
-import './version';
+export * from './SpinButton';
